+++ lib/routes/account_page.dart
@@ -27,6 +27,8 @@ import "package:flow/widgets/no_result.dart";
 import "package:flow/widgets/rates_missing_error_box.dart";
 import "package:flow/widgets/time_range_selector.dart";
 import "package:flow/widgets/transactions_date_header.dart";
+import "package:flow/widgets/transactions_selection_controller.dart";
+import "package:flow/widgets/transactions_selection_scope.dart";
 import "package:flutter/material.dart";
 import "package:go_router/go_router.dart";
 import "package:material_symbols_icons/symbols.dart";
@@ -74,12 +76,28 @@ class _AccountPageState extends State<AccountPage> {
 
   late TimeRange range;
 
+  late final TransactionsSelectionController _selection;
+
   @override
   void initState() {
     super.initState();
 
     account = ObjectBox().box<Account>().get(widget.accountId);
     range = widget.initialRange ?? TimeRange.thisMonth();
+    _selection = TransactionsSelectionController();
+    _selection.addListener(_onSelectionChanged);
+  }
+
+  @override
+  void dispose() {
+    _selection.removeListener(_onSelectionChanged);
+    _selection.dispose();
+    super.dispose();
+  }
+
+  void _onSelectionChanged() {
+    if (!mounted) return;
+    setState(() {});
   }
 
   @override
@@ -209,8 +227,11 @@ class _AccountPageState extends State<AccountPage> {
               ),
             ],
           ),
-          body: SafeArea(
-            child: switch (busy) {
+          body: TransactionsSelectionScope(
+            controller: _selection,
+            visibleTransactions: transactions ?? const [],
+            child: SafeArea(
+              child: switch (busy) {
               true => Padding(
                 padding: headerPaddingOutOfList,
                 child: Column(
@@ -231,6 +252,7 @@ class _AccountPageState extends State<AccountPage> {
               ),
               _ => GroupedTransactionsListView(
                 listType: GroupedTransactionsListViewType.reorderable,
+                selectionController: _selection,
                 mainHeader: header,
                 transactions: grouped,
                 pendingTransactions: pendingTransactionsGrouped,
@@ -254,7 +276,8 @@ class _AccountPageState extends State<AccountPage> {
                   );
                 },
               ),
-            },
+              },
+            ),
           ),
         );
       },