+++ lib/routes/home_page.dart
@@ -32,16 +32,26 @@ class HomePage extends StatefulWidget {
   const HomePage({super.key});
 
   @override
-  State<HomePage> createState() => _HomePageState();
+  State<HomePage> createState() => HomePageState();
+
+  static HomePageState of(BuildContext context) =>
+      context.findAncestorStateOfType<HomePageState>()!;
 }
 
-class _HomePageState extends State<HomePage>
+class HomePageState extends State<HomePage>
     with SingleTickerProviderStateMixin {
   late final TabController _tabController;
 
   final ScrollController _homeTabScrollController = ScrollController();
 
   late int _currentIndex;
+
+  final Map<int, bool> _hideBottomNav = {
+    0: false,
+    1: false,
+    2: false,
+    3: false,
+  };
 
   bool _navigationListenerRegistered = false;
 
@@ -125,6 +135,8 @@ class _HomePageState extends State<HomePage>
 
   @override
   Widget build(BuildContext context) {
+    final bool hideBottomNav = _hideBottomNav[_currentIndex] ?? false;
+
     return ExternalToastsHandler(
       child: CallbackShortcuts(
         bindings: {
@@ -157,18 +169,26 @@ class _HomePageState extends State<HomePage>
                   left: 0.0,
                   right: 0.0,
                   child: SafeArea(
-                    child: Frame(
-                      child: Stack(
-                        alignment: Alignment.center,
-                        children: [
-                          Navbar(
-                            onTap: (i) => _navigateTo(i),
-                            activeIndex: _currentIndex,
+                    child: IgnorePointer(
+                      ignoring: hideBottomNav,
+                      child: AnimatedSlide(
+                        offset: Offset(0.0, hideBottomNav ? 2.0 : 0.0),
+                        duration: const Duration(milliseconds: 100),
+                        child: Frame(
+                          child: Stack(
+                            alignment: Alignment.center,
+                            children: [
+                              Navbar(
+                                onTap: (i) => _navigateTo(i),
+                                activeIndex: _currentIndex,
+                              ),
+                              NewTransactionButton(
+                                onActionTap: (type) =>
+                                    _newTransactionPage(type),
+                              ),
+                            ],
                           ),
-                          NewTransactionButton(
-                            onActionTap: (type) => _newTransactionPage(type),
-                          ),
-                        ],
+                        ),
                       ),
                     ),
                   ),
@@ -259,5 +279,16 @@ class _HomePageState extends State<HomePage>
     } catch (e) {
       return false;
     }
+  }
+
+  void showNav(int index) => toggleNavVisibility(true, index);
+  void hideNav(int index) => toggleNavVisibility(false, index);
+
+  void toggleNavVisibility(bool show, int index) {
+    _hideBottomNav[index] = !show;
+
+    if (!mounted) return;
+
+    setState(() {});
   }
 }